pyproject.toml
[project]
name = "aws-strands-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "strands-agents (>=0.1.1,<0.2.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "litellm (>=1.70.0,<2.0.0)",
    "tavily-python (>=0.7.2,<0.8.0)"
]

[tool.poetry]
packages = [{include = "aws_strands_agent", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
